requirements.txt
aiosmtplib==5.0.0
alembic==1.17.1
anyio==4.11.0
bcrypt==5.0.0
cachetools==6.2.1
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
colorama==0.4.6
cryptography==46.0.3
flet==0.28.3
google-api-core==2.28.1
google-api-python-client==2.186.0
google-auth==2.41.1
google-auth-httplib2==0.2.1
google-auth-oauthlib==1.2.3
googleapis-common-protos==1.71.0
greenlet==3.2.4
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.0
httpx==0.28.1
idna==3.11
iniconfig==2.3.0
Mako==1.3.10
MarkupSafe==3.0.3
oauthlib==3.3.1
packaging==25.0
passlib==1.7.4
pillow==12.0.0
pluggy==1.6.0
proto-plus==1.26.1
protobuf==6.33.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
Pygments==2.19.2
pyparsing==3.2.5
pytest==8.4.2
python-dotenv==1.2.1
repath==0.9.0
requests==2.32.5
requests-oauthlib==2.0.0
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.44
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.5.0
